This is the initial commit of the Reddit Monitor / Reading application. I still haven't settled on a name yet. Maybe 'DReddit' as in 'dread-it', but who knows where my heart will lead me. Saturday 14th Feb, 2009, 3:39 AM and I'm sitting quetly on my own writing an application to scrape Reddit. :-(
Anyway, this is reddit.py that does that actual communication with the Reddit website and scrapes karma, comment karma and the colour of your new mail icon. It has to use cookies to get through the log in procedure but now we're there it's pretty plain sailing.

The plan is to add the following:
	A wrapper that polls this library and writes the output to terminals between commands.
	A tray icon for GNOME and KDE (together) and one for Windows XP and Windows 7.