This commit adds the tray icon feature.
When ./reddit_tray_icon.py is run it throws up a username / password prompt
and then creates a reddit alien icon the system tray. This changes to an envelope
when you have new reddit mail. It polls in the background or when you want it to.

pytrayicon.so is a component for giving you simple tray icons. It is the 64 bit version. If you want the 32 bit version, delete pytrayicon.so and rename pytrayicon32.so to pytrayicon. pytrayicon was copied straight from Gmail Notifier and is unchanged.