Merged some changes and cleaned up some issues.
I don't think it's good practice to put top level code (runs when imported or when the file is executed) at both the top and the bottom, because it's harder to track the flow of logic. There are also a few imports scattered around which needed bringing together. In the run() function I specifically set up a list of the optional features early so that we can avoid ugly error messages and writing to standard out when we don't need to. Currently tested is pynotify and xdg-open. I plan on having messages on the configuration dialog so users aren't suprised by a change in functionality but I would like it to be done in a way that doesn't scare the the user. Something like 'Hey! get pynotify for popup windows!! :-). Something cheerful so they know it's an optional feature.

Since open_url is only used to open the inbox it didn't make sense to have anything outside of the inbox class method do it.