This fixes the pynotify code making it work in all three scenarios: 	- available and on 	- available and disabled by user 	- not available
On the last commit I'd tested it without the availability of pynotify and
put in a label to tell users they can have functionality if they get it
but, as chromakode points out, it might send mac users on a bit of a wild
goose chase. I think we decide what operating systems / distros we're going
to support and try to think of a solution that works for all of them without
just having checkboxes vanish between installs.